feat(llc): add maxReconnectAttempts to WebSocketOptions
This commit introduces a new parameter `maxReconnectAttempts` to `WebSocketOptions`. This allows users to configure the maximum number of times the WebSocket will attempt to reconnect after a disconnection.

The default value is 6 attempts, which corresponds to approximately 30 seconds of reconnection attempts. If the maximum number of attempts is reached, the WebSocket will stop trying to reconnect and will disconnect.